Refractored websocket logic,finished favourites functionality
