+++ package.json
@@ -17,6 +17,7 @@
     "npm-run-all": "^4.1.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hot-toast": "^2.4.1",
     "react-redux": "^8.1.2",
     "react-router-dom": "^6.16.0",
     "react-table": "^7.8.0"
