+++ src/router/pages/details/Details.module.css
@@ -4,3 +4,18 @@
   margin: 0rem auto;
   overflow-x: scroll;
 }
+.crypto-table tr {
+  border-bottom: 1px solid black;
+}
+.crypto-table thead th {
+  padding: 0.5rem;
+}
+.crypto-table thead td:first-of-type {
+  text-align: left;
+}
+.crypto-table > tbody td {
+  padding: 0.8rem;
+}
+.crypto-table tr td:first-of-type {
+  min-width: 200px;
+}
